+++ demo/DemoV1b/Model/Base/L2Flags.generated.cs
@@ -8,13 +8,13 @@ using Newtonsoft.Json.Linq;
 
 namespace DemoV1b.Model.Base {
 /// AUTOGENERED BY caffoa ///
-    public partial class L2Flags {
+    public sealed  partial class L2Flags {
         public const string L2FlagsObjectName = "flags";
         [JsonProperty("id")]
-        public virtual string Id { get; set; }
+        public string Id { get; set; }
 
         [JsonProperty("desc")]
-        public virtual string Desc { get; set; }
+        public string Desc { get; set; }
 
         public L2Flags(){}
         public L2Flags(L2Flags other) {
@@ -22,5 +22,22 @@ namespace DemoV1b.Model.Base {
             Desc = other.Desc;
         }
         public L2Flags ToL2Flags() => new L2Flags(this);
+        public bool Equals(L2Flags other) {
+            if (ReferenceEquals(null, other)) return false;
+            if (ReferenceEquals(this, other)) return true;
+            var result = Id == other.Id && Desc == other.Desc;
+            if(result) _PartialEquals(other, ref result);
+            return result;
+        }
+        partial void _PartialEquals(L2Flags other, ref bool result);
+        public override bool Equals(object obj) => Equals(obj as L2Flags);
+        public override int GetHashCode() {
+            var hashCode = new HashCode();
+            hashCode.Add(Id);
+            hashCode.Add(Desc);
+            _PartialHashCode(ref hashCode);
+            return hashCode.ToHashCode();
+        }
+        partial void _PartialHashCode(ref HashCode hashCode);
     }
 }