+++ demo/DemoV2.Text.Json/Model/Base/STJFlags.generated.cs
@@ -8,13 +8,13 @@ using System.Text.Json.Serialization;
 
 namespace DemoV2.Text.Json.Model.Base {
 /// AUTOGENERED BY caffoa ///
-    public partial class STJFlags {
+    public sealed  partial class STJFlags {
         public const string STJFlagsObjectName = "flags";
         [JsonPropertyName("id")]
-        public virtual string Id { get; set; }
+        public string Id { get; set; }
 
         [JsonPropertyName("desc")]
-        public virtual string Desc { get; set; }
+        public string Desc { get; set; }
 
         public STJFlags(){}
         public STJFlags(STJFlags other) {
@@ -22,5 +22,22 @@ namespace DemoV2.Text.Json.Model.Base {
             Desc = other.Desc;
         }
         public STJFlags ToSTJFlags() => new STJFlags(this);
+        public bool Equals(STJFlags other) {
+            if (ReferenceEquals(null, other)) return false;
+            if (ReferenceEquals(this, other)) return true;
+            var result = Id == other.Id && Desc == other.Desc;
+            if(result) _PartialEquals(other, ref result);
+            return result;
+        }
+        partial void _PartialEquals(STJFlags other, ref bool result);
+        public override bool Equals(object obj) => Equals(obj as STJFlags);
+        public override int GetHashCode() {
+            var hashCode = new HashCode();
+            hashCode.Add(Id);
+            hashCode.Add(Desc);
+            _PartialHashCode(ref hashCode);
+            return hashCode.ToHashCode();
+        }
+        partial void _PartialHashCode(ref HashCode hashCode);
     }
 }